requirements.txt
boto3~=1.16.58
botocore~=1.19.58
python-telegram-bot~=13.1
